Rely on Django for processing form data (#6)
When an HTML form is submitted via POST the encoding was not being
respected and ASCII was assumed. In addition parsing the URL encoded
body was unnecessarily re-implemented. This step has already been taken
care of by Django and made available as request.POST.

The key change here is switching the "mimer_for_form_encoded_data" to
rely on Django's processing of the form data instead of handling it
ourselves. This resolves the encoding issue as Django properly handles
it for us.

In addition the code has been refactored for readability.